+++ tests/unit/widgets/SocialShareTest.php
@@ -21,17 +21,18 @@ class SocialShareTest extends \Codeception\Test\Unit
     const DEFAULT_CONFIG_CONFIGURATOR_ID = 'defaultConfig';
     const DEFAULT_ICONS_CONFIGURATOR_ID = 'defaultIcons';
 
-    public $vk = 'http://vk.com/share.php?url=test+url&amp;title=test+title&amp;description=test+description&amp;image=test+image+url';
-    public $facebook = 'http://www.facebook.com/sharer.php?u=test+url';
-    public $twitter = 'http://twitter.com/share?url=test+url&amp;text=test+description';
-    public $googlePlus = 'https://plusone.google.com/_/+1/confirm?hl=en&amp;url=test+url';
-    public $linkedIn = 'https://www.linkedin.com/shareArticle?mini=true&amp;url=test+url&amp;title=test+title&amp;summary=test+description';
-    public $pinterest = 'https://www.pinterest.com/pin/create/link/?url=test+url&amp;media=test+image+url&amp;description=test+description';
-    public $telegram = 'https://telegram.me/share/url?url=test+url';
-    public $viber = 'viber://forward?text=test+url';
-    public $whatsApp = 'whatsapp://send?text=test+url';
-    public $gmail = "https://mail.google.com/mail/?view=cm&amp;fs=1&amp;su=test+title&amp;body=test+description+-+test+url";
-    public $tumblr = "https://www.tumblr.com/widgets/share/tool?canonicalUrl=test+url&amp;posttype=link&amp;caption=test+title&amp;content=test+url";
+    private $vk = 'http://vk.com/share.php?url=test+url&amp;title=test+title&amp;description=test+description&amp;image=test+image+url';
+    private $facebook = 'http://www.facebook.com/sharer.php?u=test+url';
+    private $twitter = 'http://twitter.com/share?url=test+url&amp;text=test+description';
+    private $googlePlus = 'https://plusone.google.com/_/+1/confirm?hl=en&amp;url=test+url';
+    private $linkedIn = 'https://www.linkedin.com/shareArticle?mini=true&amp;url=test+url&amp;title=test+title&amp;summary=test+description';
+    private $pinterest = 'https://www.pinterest.com/pin/create/link/?url=test+url&amp;media=test+image+url&amp;description=test+description';
+    private $telegram = 'https://telegram.me/share/url?url=test+url';
+    private $viber = 'viber://forward?text=test+url';
+    private $whatsApp = 'whatsapp://send?text=test+url';
+    private $gmail = 'https://mail.google.com/mail/?view=cm&amp;fs=1&amp;su=test+title&amp;body=test+description+-+test+url';
+    private $tumblr = 'https://www.tumblr.com/widgets/share/tool?canonicalUrl=test+url&amp;posttype=link&amp;caption=test+title&amp;content=test+url';
+    private $yahoo = 'https://compose.mail.yahoo.com/?subject=test+title&amp;body=test+description+-+test+url';
 
 
     /**
@@ -41,14 +42,14 @@ class SocialShareTest extends \Codeception\Test\Unit
      *
      * @return string
      */
-    protected function getActualHTML($configurator)
+    private function getActualHTML($configurator)
     {
         $widget = new SocialShare([
             'configurator'  => $configurator,
             'url'           => 'test url',
             'title'         => 'test title',
             'description'   => 'test description',
-            'imageUrl'      => 'test image url'
+            'imageUrl'      => 'test image url',
         ]);
         ob_start();
         $widget->run();
@@ -74,6 +75,7 @@ class SocialShareTest extends \Codeception\Test\Unit
                         . "<li><a href=\"$this->whatsApp\" rel=\"noopener\" target=\"_blank\">Whats App</a></li>"
                         . "<li><a href=\"$this->gmail\" rel=\"noopener\" target=\"_blank\">Gmail</a></li>"
                         . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\">Tumblr</a></li>"
+                        . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\">Yahoo</a></li>"
                     . "</ul>"
                     . "<!--/noindex-->";
         $actualHTML = $this->getActualHTML(self::DEFAULT_CONFIG_CONFIGURATOR_ID);
@@ -97,6 +99,7 @@ class SocialShareTest extends \Codeception\Test\Unit
                 . "<li><a href=\"$this->whatsApp\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-whatsapp\"></i></a></li>"
                 . "<li><a href=\"$this->gmail\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-gmail\"></i></a></li>"
                 . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-tumblr\"></i></a></li>"
+                . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-yahoo\"></i></a></li>"
             . "</ul>"
             . "<!--/noindex-->";